BUILD : 1.3.0.74 (SVN) BUGS : N/A NOTES : Added missing library user32.lib to Makefile.inc.win32. Now Denora compiles quite fine on Windows. Included some cosmetics changes to src/users.c, provided by crazy.

+++ src/protocol/inspircd11.c
@@ -1184,7 +1184,7 @@ void inspircd_cmd_eob(void)
 int sjtoken(char *message, char delimiter, char **parv)
 {
     char *next;
-    uint16_t count;
+    unsigned count;
 
     if (!message)
         return 0;
